Added AbstractFilesTable, to support AllFilesTable
